@@ -11,84 +11,57 @@ authors = ["CTMS Reviewers <@mozilla-it/ctms-reviewers>"]
 
 [tool.poetry.dependencies]
 # These packages are mandatory and form the core of this package’s distribution.
-python = "~3.11.1"
-fastapi = "^0.84.0"
-starlette = "^0.19.1"
-requests = "^2.28.2"
-uvicorn = {extras = ["standard"], version = "^0.20.0"}
-pydantic = {extras = ["email"], version = "^1.10.7"}
-SQLAlchemy = ">1.3.23,<1.5"
-alembic = "^1.10.2"
-python-multipart = "^0.0.6"
-python-jose = {extras = ["cryptography"], version = "^3.2.0"}
-passlib = {extras = ["argon2"], version = "^1.7.4"}
-python-dateutil = "^2.8.1"
-dockerflow = "^2022.8.0"
-sentry-sdk = "^1.16.0"
-pysilverpop = "^0.2.6"
-lxml = "^4.9.2"
-prometheus-client = "0.16.0"
-structlog = "^22.3.0"
+python = "^3.12.1"
+fastapi = "^0.115.6"
+starlette = "^0.41.3"
+requests = "^2.32.3"
+pydantic = {extras = ["email"], version = "^2.10.4"}
+uvicorn = {extras = ["standard"], version = "^0.34.0"}
+SQLAlchemy = "^2.0.36"
+alembic = "^1.14.0"
+python-multipart = "^0.0.20"
+python-dateutil = "^2.9.0"
+dockerflow = {extras = ["fastapi"], version = "^2024.4.2"}
+sentry-sdk = {extras = ["fastapi"], version = "^2.19.2"}
+lxml = "^5.3.0"
+prometheus-client = "0.21.1"
 colorama = "^0.4.6"
-psycopg2 = "^2.9.5"
-google-auth = "^2.16.1"
+psycopg2 = "^2.9.10"
+google-auth = "^2.37.0"
+pydantic-settings = "^2.7.0"
+argon2-cffi = "^23.1.0"
+pyjwt = "^2.10.1"
 
 
-[tool.poetry.dev-dependencies]
-pre-commit = "^3.2.0"
-sphinx = "^6.1.3"
-pytest = "^7.2.2"
-isort = {extras = ["pyproject"], version = "^5.12.0"}
-black = "^23.1.0"
-coverage = {extras = ["toml"], version = "^7.1"}
-mypy = "^1.0"
-types-python-dateutil = "^2.8.19"
-detect-secrets = "^1.4.0"
-bandit = "^1.7.5"
-SQLAlchemy-Utils = "^0.40.0"
-pylint = "^2.17.1"
-pylint-pytest = "^1.1.2"
-types-requests = "^2.28.11"
-factory-boy = "^3.2.1"
-pytest-factoryboy = "^2.5.1"
+
+[tool.poetry.group.dev.dependencies]
+pre-commit = "^4.0.1"
+pytest = "^8.3.4"
+coverage = {extras = ["toml"], version = "^7.6"}
+mypy = "^1.14"
+types-python-dateutil = "^2.9.0"
+detect-secrets = "^1.5.0"
+bandit = "^1.8.0"
+SQLAlchemy-Utils = "^0.41.2"
+types-requests = "^2.32.0"
+factory-boy = "^3.3.1"
+pytest-factoryboy = "^2.7.0"
 backoff = "^2.2.1"
+httpx = "^0.28.1"
+ruff = "^0.8.4"
 
 [tool.pytest.ini_options]
-testpaths = [
-    "tests/unit",
+testpaths = ["tests/unit"]
+markers = [
+    "disable_autouse: disable fixtures configured with `autouse=True`",
 ]
 
-[tool.pylint]
-    [tool.pylint.'MESSAGES CONTROL']
-    disable = [
-        "R0801", # sqla has a lot of duplicate code and it is work to "fix" this
-        "C0301", # if black is happy with the line-length, let's not mess with it
-        "R0903", # this is triggered by pydantic's config and is really just a taste thing anyway
-        "R0913", # we have too-many-arguments in too many places to easily fix
-        "R0912", # Same idea as above, let's leave it since it is a matter of taste
-        "W0511", # Let us have our TODOs
-        "W0613", # sometimes unused arguments are nice for symmetry and also api methods
-        "R0914", # Leave me alone, pylint. I can have as many locals as I want
+[tool.ruff]
+target-version = "py312"
 
-        "C0114", # mozilla-it/ctms-api#100
-        "C0115", # mozilla-it/ctms-api#100
-        "C0116", # mozilla-it/ctms-api#100
-    ]
-    [tool.pylint.REPORTS]
-    output-format = 'colorized'
-    [tool.pylint.FORMAT]
-    good-names = 'e,i,f,db,SessionLocal' # Normally pylint objects to these due to not being snake_case
-    [tool.pylint.TYPECHECK]
-    generated-members='alembic.*'
-    [tool.pylint.MASTER]
-    extension-pkg-whitelist='pydantic'
-    load-plugins='pylint_pytest'
-    ignore='third_party'
-    ignore-patterns = "migrations/.*" # Not worth fixing yet
-
-[tool.isort]
-profile = "black"
-
+[tool.ruff.lint]
+select = ["PL", "I"]
+ignore = ["PLR2004", "PLR0913"]
 
 [tool.coverage]
 # https://github.com/nedbat/coveragepy
