@@ -1,10 +1,15 @@
-from datetime import datetime, timezone
-from typing import Optional
+from __future__ import annotations
 
-from pydantic import UUID4, AnyUrl, Field, root_validator
+from typing import TYPE_CHECKING, Optional, Union
+
+from pydantic import UUID4, ConfigDict, Field, model_validator
 
 from .base import ComparableBase
+from .common import AnyUrlString, ZeroOffsetDatetime
 from .email import EMAIL_ID_DESCRIPTION, EMAIL_ID_EXAMPLE
+
+if TYPE_CHECKING:
+    from .contact import ContactInBase, ContactPatchSchema
 
 
 class WaitlistBase(ComparableBase):
@@ -13,212 +18,139 @@ class WaitlistBase(ComparableBase):
 
     This is meant to serve as the common and generic schemas to
     all waitlists.
-
-    In this implementation phase, it cohabits with individual (non-generic)
-    schemas of Relay and VPN.
-
-    TODO waitlist: once Basket leverages the `waitlists` field, we can drop
-    `RelayWaitlistBase` and `VpnWaitlistBase`.
     """
 
     name: str = Field(
         min_length=1,
         description="Basket slug for the waitlist",
-        example="new-product",
+        examples=["new-product"],
     )
-    source: Optional[AnyUrl] = Field(
+    source: Optional[AnyUrlString] = Field(
         default=None,
         description="Source URL of subscription",
-        example="https://www.mozilla.org/en-US/",
+        examples=["https://www.mozilla.org/en-US/"],
     )
     fields: dict = Field(
-        default={}, description="Additional fields", example='{"platform": "linux"}'
+        default={}, description="Additional fields", examples=['{"platform": "linux"}']
+    )
+    subscribed: bool = Field(
+        default=True, description="True to subscribe, False to unsubscribe"
+    )
+    unsub_reason: Optional[str] = Field(
+        default=None, description="Reason for unsubscribing"
     )
 
     def __lt__(self, other):
         return self.name < other.name
 
-    @root_validator
-    def check_fields(cls, values):  # pylint:disable = no-self-argument
-        if "name" in values:
-            validate_waitlist_fields(values["name"], values.get("fields", {}))
-        return values
+    @model_validator(mode="after")
+    def check_fields(self):
+        """
+        Once waitlists will have been migrated to a full N-N relationship,
+        this will be the only remaining VPN specific piece of code.
+        """
+        if self.name == "relay":
 
-    class Config:
-        orm_mode = True
+            class RelayFieldsSchema(ComparableBase):
+                geo: Optional[str] = CountryField()
+                model_config = ConfigDict(extra="forbid")
+
+            RelayFieldsSchema(**self.fields)
+
+        elif self.name == "vpn":
+
+            class VPNFieldsSchema(ComparableBase):
+                geo: Optional[str] = CountryField()
+                platform: Optional[str] = PlatformField()
+                model_config = ConfigDict(extra="forbid")
+
+            VPNFieldsSchema(**self.fields)
+
+        else:
+            # Default schema for any waitlist.
+            # Only the known fields are validated. Any extra field would
+            # be accepted as is.
+            # This should allow us to onboard most waitlists without specific
+            # code change and service redeployment.
+            class DefaultFieldsSchema(ComparableBase):
+                geo: Optional[str] = CountryField()
+                platform: Optional[str] = PlatformField()
+
+            DefaultFieldsSchema(**self.fields)
+
+        return self
+
+    model_config = ConfigDict(from_attributes=True)
 
 
 # No need to change anything, just extend if you want to
 WaitlistSchema = WaitlistBase
 
 
-class WaitlistInSchema(WaitlistBase):
-    """Schema for input data."""
+WaitlistInSchema = WaitlistBase
 
-    subscribed: bool = Field(
-        default=True, description="True to subscribe, False to unsubscribe"
+
+class WaitlistTimestampedSchema(WaitlistBase):
+    create_timestamp: ZeroOffsetDatetime = Field(
+        description="Waitlist data creation timestamp",
+        examples=["2020-12-05T19:21:50.908000+00:00"],
     )
-
-    @root_validator
-    def check_fields(cls, values):  # pylint:disable = no-self-argument
-        if "subscribed" in values and values["subscribed"]:
-            return super().check_fields(values)
-        # If subscribed is False, we don't need to validate fields.
-        return values
-
-    def orm_dict(self):
-        """TODO: is there a native way to exclude attrs for ORM?"""
-        dict_for_orm = self.dict()
-        del dict_for_orm["subscribed"]
-        return dict_for_orm
-
-
-class UpdatedWaitlistInSchema(WaitlistInSchema):
-    update_timestamp: datetime = Field(
-        default_factory=lambda: datetime.now(timezone.utc),
+    update_timestamp: ZeroOffsetDatetime = Field(
         description="Waitlist data update timestamp",
-        example="2021-01-28T21:26:57.511Z",
+        examples=["2021-02-04T15:36:57.511000+00:00"],
     )
 
 
-class WaitlistTableSchema(WaitlistBase):
+class WaitlistTableSchema(WaitlistTimestampedSchema):
     email_id: UUID4 = Field(
         description=EMAIL_ID_DESCRIPTION,
-        example=EMAIL_ID_EXAMPLE,
+        examples=[EMAIL_ID_EXAMPLE],
     )
-    create_timestamp: datetime = Field(
-        description="Waitlist data creation timestamp",
-        example="2020-12-05T19:21:50.908000+00:00",
+    model_config = ConfigDict(extra="forbid")
+
+
+def CountryField():
+    return Field(
+        default=None,
+        max_length=100,
+        description="Waitlist country",
+        examples=["fr"],
     )
-    update_timestamp: datetime = Field(
-        description="Waitlist data update timestamp",
-        example="2021-02-04T15:36:57.511000+00:00",
+
+
+def PlatformField():
+    return Field(
+        default=None,
+        max_length=100,
+        description="VPN waitlist platforms as comma-separated list",
+        examples=["ios,mac"],
     )
 
-    class Config:
-        extra = "forbid"
 
-
-def validate_waitlist_fields(name: str, fields: dict):
+class RelayWaitlistSchema(ComparableBase):
     """
-    Once waitlists will have been migrated to a full N-N relationship,
-    this will be the only remaining VPN specific piece of code.
-    """
-    if name == "relay":
-
-        class RelayFieldsSchema(ComparableBase):
-            geo: str = Field(
-                max_length=100,
-                description="Waitlist country",
-                example="fr",
-            )
-
-            class Config:
-                extra = "forbid"
-
-        RelayFieldsSchema(**fields)
-
-    elif name == "vpn":
-
-        class VPNFieldsSchema(ComparableBase):
-            geo: str = Field(
-                max_length=100,
-                description="Waitlist country",
-                example="fr",
-            )
-            platform: Optional[str] = Field(
-                default=None,
-                max_length=100,
-                description="VPN waitlist platforms as comma-separated list",
-                example="ios,mac",
-            )
-
-            class Config:
-                extra = "forbid"
-
-        VPNFieldsSchema(**fields)
-
-
-def validate_waitlist_newsletters(values):
-    """
-    This helper validates that when subscribing to `relay-*-waitlist`
-    newsletters, the country is provided.
-    # TODO waitlist: remove once Basket leverages the `waitlists` field.
-    """
-    if "newsletters" not in values:
-        return values
-
-    newsletters = values["newsletters"]
-    if not isinstance(newsletters, list):
-        return values
-
-    relay_newsletter_found = False
-    for newsletter in newsletters:
-        if newsletter.subscribed and newsletter.name.startswith("relay-"):
-            relay_newsletter_found = True
-            break
-
-    if not relay_newsletter_found:
-        return values
-
-    # If specified using the legacy `relay_waitlist`
-    relay_country = None
-    relay_waitlist = values.get("relay_waitlist")
-    if relay_waitlist:
-        relay_country = relay_waitlist.geo
-    elif "waitlists" in values:
-        # If specified using the `waitlists` field (unlikely, but in our tests we do)
-        waitlists = values["waitlists"]
-        if isinstance(waitlists, list):
-            for waitlist in waitlists:
-                if waitlist.name == "relay":
-                    relay_country = waitlist.fields.get("geo")
-
-    # Relay country not specified, check if a relay newsletter is being subscribed.
-    if not relay_country:
-        raise ValueError("Relay country missing")
-
-    return values
-
-
-class RelayWaitlistBase(ComparableBase):
-    """
-    The Mozilla Relay Waitlist schema.
-
-    TODO waitlist: remove once Basket leverages the `waitlists` field.
+    The Mozilla Relay Waitlist schema for the read-only `relay_waitlist` field.
     """
 
     geo: Optional[str] = Field(
         default=None,
         max_length=100,
         description="Relay waitlist country",
-        example="fr",
+        examples=["fr"],
     )
-
-    class Config:
-        orm_mode = True
+    model_config = ConfigDict(from_attributes=True)
 
 
-# No need to change anything, just extend if you want to
-RelayWaitlistInSchema = RelayWaitlistBase
-RelayWaitlistSchema = RelayWaitlistBase
-
-
-class VpnWaitlistBase(ComparableBase):
+class VpnWaitlistSchema(ComparableBase):
     """
-    The Mozilla VPN Waitlist schema.
-
-    This was previously the Firefox Private Network (fpn) waitlist data,
-    with a similar purpose.
-
-    TODO waitlist: remove once Basket leverages the `waitlists` field.
+    The Mozilla VPN Waitlist schema for the read-only `vpn_waitlist` field
     """
 
     geo: Optional[str] = Field(
         default=None,
         max_length=100,
         description="VPN waitlist country, FPN_Waitlist_Geo__c in Salesforce",
-        example="fr",
+        examples=["fr"],
     )
     platform: Optional[str] = Field(
         default=None,
@@ -227,13 +159,6 @@ class VpnWaitlistBase(ComparableBase):
             "VPN waitlist platforms as comma-separated list,"
             " FPN_Waitlist_Platform__c in Salesforce"
         ),
-        example="ios,mac",
+        examples=["ios,mac"],
     )
-
-    class Config:
-        orm_mode = True
-
-
-# No need to change anything, just extend if you want to
-VpnWaitlistInSchema = VpnWaitlistBase
-VpnWaitlistSchema = VpnWaitlistBase
+    model_config = ConfigDict(from_attributes=True)
