@@ -24,51 +24,8 @@ from .fxa import (
     UpdatedFirefoxAccountsInSchema,
 )
 from .mofo import MozillaFoundationInSchema, MozillaFoundationSchema
-from .newsletter import NewsletterInSchema, NewsletterSchema, UpdatedNewsletterInSchema
-from .product import ProductBaseSchema
-from .stripe_customer import (
-    StripeCustomerCreateSchema,
-    StripeCustomerModelSchema,
-    StripeCustomerOutputSchema,
-    StripeCustomerUpsertSchema,
-)
-from .stripe_invoice import (
-    StripeInvoiceCreateSchema,
-    StripeInvoiceModelSchema,
-    StripeInvoiceOutputSchema,
-    StripeInvoiceUpsertSchema,
-)
-from .stripe_invoice_line_item import (
-    StripeInvoiceLineItemCreateSchema,
-    StripeInvoiceLineItemModelSchema,
-    StripeInvoiceLineItemOutputSchema,
-    StripeInvoiceLineItemUpsertSchema,
-)
-from .stripe_price import (
-    StripePriceCreateSchema,
-    StripePriceModelSchema,
-    StripePriceOutputSchema,
-    StripePriceUpsertSchema,
-)
-from .stripe_subscription import (
-    StripeSubscriptionCreateSchema,
-    StripeSubscriptionModelSchema,
-    StripeSubscriptionOutputSchema,
-    StripeSubscriptionUpsertSchema,
-)
-from .stripe_subscription_item import (
-    StripeSubscriptionItemCreateSchema,
-    StripeSubscriptionItemModelSchema,
-    StripeSubscriptionItemOutputSchema,
-    StripeSubscriptionItemUpsertSchema,
-)
-from .waitlist import (
-    RelayWaitlistInSchema,
-    UpdatedWaitlistInSchema,
-    VpnWaitlistInSchema,
-    WaitlistInSchema,
-    WaitlistSchema,
-)
+from .newsletter import NewsletterInSchema, NewsletterSchema, NewsletterTableSchema
+from .waitlist import WaitlistInSchema, WaitlistSchema, WaitlistTableSchema
 from .web import (
     BadRequestResponse,
     NotFoundResponse,
